requirements.txt
azure-devops>=7.1.0b3
msrest>=0.7.1
openpyxl>=3.1.0
python-dotenv>=1.0.0
python-dateutil>=2.8.0
